requirements.txt
streamlit
pandas
requests
python-dotenv
tqdm
sentence-transformers
torch
